Window planning correctly gets params from agg and window subplans
During window planning, when making a nested loop wrapper plan for the
window and aggregate plans, we make the right and left sides of the
nested loop into SubqueryScan and then set their plans to be the
subplans of their respective SubqueryScans. Previously, we did not
correctly recurse through these newly created subplans (formerly the agg
and window plans) and ensure that the extParam and allParam are
correctly bubbled up. That meant that later, when we attempt to do this
in SS_finalize_plan, it was too late, because we do not recurse into
SubqueryScan nodes and bubble up those params and instead simply grab
the top-level subplan's extParam, which, now, would not have been set
correctly.

This commit resolves that issue as well as adds several tests which
assert on the plan tree and when params are expected to be set.

We've added an extension that acts as an integration test for the
planner. We build an extension that exposes a function that plans a
query string and inspects the output PlannedStmt. This style test
should help us define expectations about what happens during planning
from the perspective of the executor or other components dependent on
the planner. This approach was inspired by `src/test/modules/pred_test`
upstream.