Attempt to add ATTACH PARTITION syntax to gram.y
This doesn't compile. I was attempting to add ATTACH and have it just
use the same code as when we make an AlterPartitionCmd with
`alter_table_partition_cmd` but I don't know bison, so, the $$ = $1 is
wrong. I thought if I could somehow set the command subtype before going
into the code for the alter_table_partition_cmd token, then I could
modify that code minimally to not reset the subtype. However, I have no
idea how to get this to go to the code that alter_table_partition_cmd
token goes to (I thought that by using the argument in position 1 [$1] I
would auto-magically go there since that is what it looks like
alter_table_partition_cmd does)