Make ATTACH PARTITION add to pg_partition_rule correctly
By moving the addition of the new leaf partition to the catalog to
earlier (i.e. during parsing/transformation as oppposed to during
execution) we are able to benefit from the transformations done in
`transformAlterTable_all_PartitionStmt`. Basically, we will be creating
a PartitionBy data structure before returning from parsing. Now, if you
select from pg_partition_rule, you will see the correct boundaries for
this new leaf partition.

yay! it works...but, you can't actually use this partition which was
attached because the whole partition table created with the existing
partition syntax has all of its partitions defined and so there is
nowhere for the newly attached partition to be.

Also, we will need to add logic that either uses (and validates) an
existing constraint on the table or adds one from the `FOR VALUES`
clause

Signed-off-by: Shoaib Lari <slari@pivotal.io>