Update parser to support new root partition table creation DDL
Previously, you needed to do this

```
CREATE TABLE t1 (
    a integer,
    b integer
) DISTRIBUTED BY (a) PARTITION BY RANGE(b)
          (
          START (1) END (2) EVERY (1)
          );
```
to make a partition table.

If we want to attach a leaf partition to a root and have it actually be
used, we need to allow creation of a "root only" partition table.
To accomplish this, we added to the parser a new keyword, ROOTONLY.

Proposed syntax:

```
CREATE TABLE part (
    a integer,
    b integer
) DISTRIBUTED BY (a) PARTITION BY ROOTONLY RANGE (b) ;
```

Currently, this compiles, but when you attempt to use the new syntax,
you get an error during the transformation step which validates the
partition boundaries. (See TODO for details).

Signed-off-by: Shoaib Lari <slari@pivotal.io>