Resolve FIXME for validatepart by passing relpersistence of root
MergeAttributes is used for defining a relation and is used here, likely
as a convenience as it seems like the wrong function for the job, simply
to get the column definitions for the partition table which we are
considering adding a partition to as a child. We only hit this code path
currently when adding a partition as part of SPLIT PARTITION or ADD
PARTITION, both of which are making a new table and adding a new entry
to pg_class for this additional partition. That means that, for now, we
would never have a temporary relation to which we are adding a new child
since you cannot create a temporary partition table and cannot specify
temporary as a keyword when splitting a partition.

However, there is nothing inherent in this function that should require
it is only used in those circumstances and it seems wrong to make
atpxPart_validate_spec aware of the intentions of the caller.

Previously this function simply hard-coded in false for the relation
persistence since the parameter was simply `isTemp`. Once the options
for relation persistence were expanded to included unlogged, this
paramter was changed to take a relpersistence. In MergeAttributes, for
the part which we actually hit when calling it from here (we pass in the
schema as NIL and therefore hit only half of the MergeAttributes code)
the `supers` parameter is actually that of the parent partition and
includes relpersistence, so, by passing in the relpersistence of the
parent as relpersistence here, the checks we do around relpersistence are
redundant because we are comparing the parent's relpersistence to its
own. However, because, currently, this function is only called when we
are making a new relation that, because we don't allow a different
persistence to be specified for the child would actually just be using
the relpersistence of the parent anyway, by passing it in hard-coded we
would actually be incorrectly assuming that we are creating a permanent
relation always.

Since MergeAttributes was overkill anyway, we wrote a new helper
function, SetSchemaAndConstraints, to get the schema and constraints of
a relation. This function doesn't do very many special validation checks
that may be required by callers when using it in the context of
partition tables (so user beware), however, it is probably only useful
in the context of partition tables because it assumes constraints will
be cooked, which, wouldn't be the case for all relations.
We split it into two smaller inline functions for clarity. We also felt
this would be a useful helper function in general, so we extern'd it.

I am not adding a test here, because, with the code as it is now,
nothing would break if we kept it as permanent. It would never end up
being temporary and, with because we don't use this relpersistence to
set that of the child and also we don't do any validation checks about
unlogged in MergeAttributes, we would have ended up with the same
effective relpersistence for the child either way.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>
Co-authored-by: Melanie Plageman <mplageman@pivotal.io>