Attempt to handle param messages to sender
Make a new flag and attempt to handle messages sent to sender from
receiver which contain params. It turns out I have no idea how to do
this or how this code works...
See the commit following this one to understand what this is supposed to
do and why we want it.