Make rescanmotion send something from receiver to sender
So, I want a new message which has a param to be sent by the receiver to
the sender. I tried making one in a crazy way with some hard-coded data.
Then I want handleMsgParams to check interconnect thread (I didn't get
that work), so instead set some hardcoded stuff in the expression
context of the indexonlyscanstate planstate.