Sender should poll for params
So, we decided that the main thread in the sender process will poll
after sending an EOS message to the receiver to check for new params.
This makes it synchronous. It will process the param, put it in the
econtext, and then initiate a rescan.

I tried adding this here, but now the sender is stuck polling and it
seems like the receiver is checking for data and is waiting. I'm not
sure why.