Add variable to communicate between IC and main thread
Add a variable which is updated by main thread to let IC thread know
that it can EOS ACK (the receiver can). Do this once we've run out of
parameters. While we still have parameters, don't send acks.

I'm afraid that this will have a problem with timeout of waiting for EOS
Ack.