Add cmockery test of zstd compression module
This commit adds basic unittesting of zstd compression module.
The convenience macro EXPECT_ELOG which was previously in
src/backend/utils/mmgr/test/runaway_cleaner_test.c
was of use when mocking elog, so add it to a neutral location. Since it
uses all of the CMockery functions, add it in the existing cmockery
directory.

Mock out ZSTD by hand. elog is auto-mocked with the machinery in
$MOCK_DIR

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>
Co-authored-by: David Kimura <dkimura@pivotal.io>