Rig the members of conn_info
So, for some reason, src and dest are mixed up in the MotionConn's
conn_info (the packet header). I thought manually changing them in
sendparam would help.
I think it made a change to the behavior.
Either way, the behavior now is that the receiver keeps trying to send
param packets forever and the sender exits and is idle.