Non-working question sender sending EOS for param
So, as we discussed, a sender should send EOS to the receiver who sent
it the param for which it is indicating EOS.
The idea is to use the `extraSeq` field in the packet.
The places I imagined having to add code for this to work are:
1) receiver sets pkt->extraSeq to something identifying it before
sending an ack with a param
2) sender saves pkt->extraSeq in the param context (in a new field)
3) sender checks if the connection which it is sending the EOS to is for
the receiver which is identified by the param in the pkt it is going to
send

So, this doesn't seem to work for a couple of reasons:
First, I recognize that the contentId of the receiver is not enough to
identify it, however, I thought it was a start.

Second, I am logging the values of the pkt extraSeq and the paramContext
value for paramSeq and they are never the same.
I checked the other places that extraSeq is modified, and maybe that is
the problem, however, it seems like the paramContext is not set
correctly either.
It seems also like the dstContentId in the connection doesn't match what
it should?
I thought that by adding a number in the ExprContext
data structure which is referred to in the estate that that would
actually be available from any planstate node which initialized its
expression context to point to the estate expression context and that by
setting it in the paramcontext of a given plan state, I was setting it
in the estate. That might be wrong.

Also, this logic, once correct, should be added to doSendTuple.
In order for this to work on a per tuple basis, we can maybe use
`es_per_tuple_exprcontext`