Try and get EOS sent where it should go: take 2
So, this commit is a combination of a patch Gang wrote and a thing I was
trying to do to use the connection itself to manage the issue of retries
and use the EOS packet to synchronize the param packet
sending without using a static int inside of sendParam.

I didn't have time to actually make the idea work/I'm unconvinced it's
actually a good idea.

This is what I did:

Initialize a field in the MotionConn which indicates the sequence of the
last param which the sender has sent EOS for in setup for each receiver
conn.
For the receiver, when it receives the EOS, it will check the EOS packet's
paramseq and contentid.
If the EOS packet's paramseq == the paramEosSeq in the connection, then
that is EOS for the last param it sent, so it is time to send a new
param. Set nextParamSeq to paramEosSeq++
If the EOS packet's paramseq < connection's paramEosSeq, then it is a
retry of the EOS, so do nothing.
If the EOS packet's paramseq > connection's paramEosSeq, I guess that
means we got an EOS for a param we haven't sent yet?? I don't know what
that means.

Then, in sendParam, set the paramSeq to the connection's nextParamSeq.

The sender would then get a param in handleDataPacket and if the
paramSeq is > than the last param it has sent EOS for, set it in
paramcontext.

Also, when the sender sends EOS, it would use the paramseq from its
paramcontext to set the last EOS'd param seq field in the connection.

Note: the above may not make sense;) I left the static paramSeq for now
because the thing I tried didn't work.

Also, what about where the sender sends the actual tuples (not just
EOS)?

Co-authored-by: Gang Xiong <gxiong@pivotal.io>