Update receiver logic
So, I added some variables to the ICGlobalControlInfo to communicate
between receiver and IC about if we should send EOS Ack or not/if we
have more params.
NLJ will set this for the IC thread to let it know that we don't have
more params.
So, I think IC thread needs to know if it is an EOS packet before
deciding whether or not to send the EOS Ack to it or not
before
I added send_eos_ack so that the NLJ could indicate that it has no more
params for the IC thread.
I added lastParameterValue so taht the rescan motion receiver could
block until the IC thread has sent the param, because otherwise we could
end up updating the param several times before the IC has time to send
the params.

Also, I deleted the (I think) dead code in execUnsortedReceiver that
sent a chunk.