Let NLJ path inform motion child it will be rescanned
Remove the explicit check for indexonlyscan child in ExecInitMotion so
that, for a path for a NLJ plan with a motion on its inner side, the
CdbMotionPath sets a flag that it will do a rescan and that flag
is used to populate the Motion plan node when creating the plan. This
can be referenced during ExecInitNode and lets us set up the
paramsequence which is used for, amongst other things, flow control for
doing the rescan.

Also, remove some if0'd code