Attempt to handle having to pass a param through two motions
Lots of super hack-y walker code here to try and handle the case where
the slice on the outer side of the NLJ does not have a leaf which is a
scan. For a query like this:

CREATE TABLE TEXT_TBL (f1 text);
INSERT INTO TEXT_TBL VALUES ('doh!');
INSERT INTO TEXT_TBL VALUES ('hi de ho neighbor');
CREATE TABLE INT8_TBL(q1 int8, q2 int8);
INSERT INTO INT8_TBL VALUES('  123   ','  456');
INSERT INTO INT8_TBL VALUES('123   ','4567890123456789');
INSERT INTO INT8_TBL VALUES('4567890123456789','123');
INSERT INTO INT8_TBL VALUES(+4567890123456789,'4567890123456789');
INSERT INTO INT8_TBL VALUES('+4567890123456789','-4567890123456789');

postgres=# explain (costs off) select * from
text_tbl t1
left join int8_tbl i8
on true,
lateral (select i8.q1, t2.f1 from text_tbl t2 limit 1) as ss
where t1.f1 = ss.f1;
		       QUERY PLAN
------------------------------------------------------------------
Gather Motion 3:1  (slice4; segments: 3)
->  Nested Loop
    Join Filter: (t1.f1 = t2.f1)
    ->  Nested Loop Left Join
	  ->  Seq Scan on text_tbl t1
	  ->  Broadcast Motion 3:3  (slice1; segments: 3)
		->  Seq Scan on int8_tbl i8
    ->  Redistribute Motion 1:3  (slice3; segments: 1)
	  Hash Key: t2.f1
	  ->  Limit
		->  Gather Motion 3:1  (slice2; segments: 3)
		      ->  Limit
			    ->  Seq Scan on text_tbl t2
Optimizer: Postgres query optimizer
(14 rows)

the redistribute sender in slice 3 needs to pass i8.q1 to slice 2, where
it will actually be used in the scan.
Previously, execMotionSender expected there to be a scan as the leaf
node -- it also assumed that there would be no intermediate nodes.
Walking over the intermediate nodes is fine, however, we now have to
handle the case of slice3. It is a Redist motion sender which will
receive the param from slice4, the redist motion receiver will send that
param. Slice 3 then needs to get the param set in the MotionState node
of the Gather Receiver which is the leaf node for its slice.

Then the Gather Receiver needs to send this param to the Gather sender
in slice 2 so that it can then initialize the expression context of scan
of t2 with the param.

I tried adding all manner of walker in order to set the parameter member
of the leaf motionstate node in the case of that kind of slice and set
the expression context in the case of the slice ending in a scan.

However, it now hangs and does not correctly pass the param.

I need to debug more, but, it might have something to do with an EOS ack
not getting sent back to the Gather receiver? Or maybe something to do
with the NLJ indicating done to the interconnect thread in the
globalinfostruct but that not getting sent to slice2.
Not sure.