Make UDAs that leak and OOM
leaky_sum() is an aggregate whose combine function is defined to
construct and aggstate which is quite large and allocate that memory
with palloc, so it is not accounted for in mpool.

good_sum() is an aggregate whose combine function is also defined to
construct and aggstate which is quite large but which allocates that
memory with mpool_alloc(), so, that memory is considered when deciding
to spill.

We have patched the constructors used by all of the combine functions
currently in the catalog, however, a user is free to define an
aggregate, and, if she uses palloc() to allocate memory, this memory
will go unaccounted for and could result in violating
gp_vmem_protect_limit.

We made good_sum() instead of using regular sum() because the larger
transvalue allows us to demonstrate the spill vs OOM with a much smaller
number of tuples.

Note that gpdiff doesn't understand the `(ANALYZE, COSTS OFF)` syntax
with parentheses, so the test won't pass.

To see the example in action, do the following:

gpconfig -c gp_vmem_protect_limit -v 256

then in src/test/regress:

./pg_regress --dlpath . make_it_leak

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>