Avoid ADD COLUMN full table rewrite for AOCO child partitions
ALTER TABLE ADD COLUMN to a child partition with storage type AOCO
should not trigger a full table rewrite. Instead, only data
corresponding to the new column should be written.

There was a regression caused by an implementation detail of the ALTER
TABLE machinery. Child partition handling relied on the
AlteredTableInfo->subcmds in order to determine eligibility for the
avoid full table rewrite optimization.
For ALTER TABLE ADD COLUMN, the subcmds are not populated for child
partitions. subcmds are not populated because the "prep" phase of ALTER
TABLE ADD COLUMN does not recurse into child partitions. Instead, the
recursion is done during the "exec" phase (ATExecAddColumn()).

However, other subcommand types (e.g. ALTER TABLE ALTER COLUMN TYPE) do
do recursion during the "prep" phase. This means that the subcmds could
be populated from other subcommand types (you can have multiple
subcommands in a given query). Therefore, we could not easily populate
the subcommands for ALTER TABLE ADD COLUMN during the "exec" phase.

So, we decoupled the logic for the table rewrite optimization from the
contents of the subcmd.  Based on the subcmds of the parent partition,
we determine if the optimization is *possible* then use the storage type
of the child partitions to make a final decision as to whether or not
the full table rewrite is required on a per-partition basis.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>