Set AOCO ADD COL flag on all relations up front
Previous commit tried to determine eligibility for all partitions using
the root partition by checking if the current AlteredTableInfo had a
newvals list. It turns out that this condition is true for all
partitions in a partition hierarchy, so all partitions were getting this
flag set in the first if statement.

We didn't notice that this doesn't
work because for most other AT subcommand types (e.g. drop column and
alter type), they actually populate their array member in the subcmds
array even for child partitions, so the flag was getting correctly not
set (it should not be set when we mix types of subcommands). The other
time we should do a full table rewrite is when the storage type is not
AOCO, and, since we always guard against that.
So, we never incorrectly didn't rewrite.

However, we would incorrectly do the full table rewrite for a
subpartitioned table's leaf partitions when they were eligible.
TODO: add a test for this

Using a local variable for communicating from the root to children
doesn't work for 3+ level partition table unless you set the flag on all
descendants proactively.

That is what we did.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>