Enable compiler checking printf-like functions in Orca
Postgres defines the macro `pg_attribute_printf` to be used with
printf-style functions in src/include/c.h for argument type checking.

`pg_attribute_printf` uses the format function attribute, however, the
format archetype differs from compiler to compiler.

Postgres defines another macro, `PG_PRINTF_ATTRIBUTE`, for the format
archetype in pg_config.h, a generated file. It is defined to be
`gnu_printf` when compiling with gcc or `printf` otherwise.

When Orca is built independently with cmake, it may not have the
generated pg_config.h.
So, create an analagous method for Orca to enable checking of
printf-like functions by compilers supporting it.

Set the format archetype conditionally based on compiler in a new macro,
`GPOS_FORMAT_ARCHETYPE`, defined in a generated file gpos/config.h.

Add a new file, attributes.h, for defining macros in Orca relating to
compiler attributes and define a macro in it, `GPOS_ATTRIBUTE_PRINTF`, which
uses the format archetype defined in gpos/config.h.

When building with gnu-make, `GPOS_FORMAT_ARCHETYPE` will be defined as
`PG_PRINTF_ATTRIBUTE`, since pg_config.h is assumed to exist.

This commit also uses the new macro in a few functions which would
benefit from it.

Co-authored-by: Melanie Plageman <melanieplageman@gmail.com>