Remove bogus calls to CStringStatic::AppendFormat.
In commit 942448b7a4bfa we temporarily suppressed a compiler warning
that printf-like functions are missing the __format__ attribute for
consistency checking. In order to reinstate such diagnostics, we will
add format attributes to various GPOS methods. Once that happens (in a
forthcoming commit), the compiler catches a call site that is passing a
nonliteral format string to a printf-like function:

../libgpos/server/src/unittest/gpos/io/CFileTest.cpp:285:34: error: format not a string literal and no format arguments [-Werror=format-security]
  strTmpFile.AppendFormat(szTmpDir);

At first we were tempted to merely replace AppendFormat with
AppendBuffer. After all that looks like the most logical substitute. But
upon closer inspection, the whole usage pattern in
CFileTest::Unittest_MkTmpFile is quite bogus. In addition, AppendBuffer
has a truncating behavior that might be more surprising.

In addition to fixing the warning, this patch rearranges the code a bit,
adds comments to make sure the next person reading it pulls less of
their hair out.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>